% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_colors_to_darkmode.R
\name{adjust_colors_to_darkmode}
\alias{adjust_colors_to_darkmode}
\title{Adjust colors based on lightness}
\usage{
adjust_colors_to_darkmode(colors, threshold = c(30, 70), amount = 0.2)
}
\arguments{
\item{colors, }{character vector of colors, e.g. hex codes}

\item{threshold, }{numeric, if the L is below the first value, it gets lightened, if above the second value, it gets darkened. It's possible to pass one value only.}

\item{amount, }{numeric, how much lighter/darker. Default 0.15}
}
\value{
a character of the same length as colors with the (potentially) modified values.
}
\description{
We may change the colors when entering dark mode. If the color is too dark, we make it a bit lighter. If the color is bright, we make it a bit darker. 
We use L from the HCL colorspace to determine the lightness/darkness.
We use the colorspace::lighten function.
}
