% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_pal.R
\name{bs_pal}
\alias{bs_pal}
\title{Bootstrap Color Palette}
\usage{
bs_pal(dark = FALSE)
}
\arguments{
\item{dark}{logical, default FALSE}
}
\value{
a named character vector, with names body-bg, body-color, body-emphasis-color, body-secondary-color, body-secondary-bg,
body-tertiary-color, body-tertiary-bg, headings-color, link-color, link-hover-color, 
code-color, highlight-color, highlight-bg, border-color, border-color-translucent, 
form-valid-color, form-valid-border-color, form-invalid-color, form-invalid-border-color
}
\description{
This function uses bslib::bs_current_theme() and bslib::bs_get_variables() to query the root level colors. 
If dark ist TRUE, it returns the "-dark" variables. The names of the returned vector are however those used for the light mode.
}
\details{
Note that there is a slightly different naming convention for highlight/highlight-bg. In dark mode, these colors are stored under
mark-color-dark/mark-bg-dark.

This function should be called in a reactive context.

See https://getbootstrap.com/docs/5.3/customize/color-modes/ for more information on the color mode.
}
