% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormode_srv.R
\name{colormode_srv}
\alias{colormode_srv}
\title{colormode_ui/srv is a shiny module for managing colors; in particular enabling dark mode.}
\usage{
colormode_srv(id = "colormode", r, verbose = FALSE)
}
\arguments{
\item{id}{character, shiny id. Default "colormode"}

\item{r}{shiny::reactiveValues object}

\item{verbose}{logical, diagnostic message, default FALSE}
}
\value{
used for its side effects
}
\description{
The server follows the "petite r" approach. It expects a reactiveValues parameter r. It modifies "colormode" entries
}
\details{
To save preferences across sessions, the lstore module is used if it is available.

The module follows a singleton design pattern, hence the id is preset to "colormode". It is strongly recommended to keep that id.

See shiny::runApp(system.file("examples/01_colormode", package="shiny.gems"))
to see the module in action, see colormode_srv for implementation details.
}
