% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormode_ui.R
\name{colormode_ui}
\alias{colormode_ui}
\title{colormode_ui/srv is a shiny module for managing colors; in particular enabling dark mode.}
\usage{
colormode_ui(id = "colormode", ...)
}
\arguments{
\item{id}{character, shiny id. Default "colormode"}

\item{...}{further arguments that are passed to shiny::radioButtons(). In particular, "width" and "inline" can be set this way.}
}
\value{
a shiny::div
}
\description{
The UI produces a subform that can be integrated in a settings/prefences tab. Inspired by the wikipedia mobile version (Feb. 2025), it displays a radiobutton choice
between "light", "dark", and "automatic". The default is "light".
}
\details{
For the automatic setting, Javascript is used to determine the local hour. The Javascript code curates a variable "auto_status" that is accessible tin the server
module. In particular, the Javascript updates "auto_status" when at 8pm and 6am.

The module follows a singleton design pattern, hence the id is preset to "colormode". It is strongly recommended to keep that id.

Run shiny::runApp(system.file("examples/01_colormode", package="shiny.gems"))
to see the module in action, see colormode_srv for implementation details.
}
