% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts_srv.R
\name{fonts_srv}
\alias{fonts_srv}
\title{fonts_ui/srv is a shiny module for managing fonts and fonts size}
\usage{
fonts_srv(id = "fonts", r, verbose = TRUE)
}
\arguments{
\item{id}{character, shiny id. Default "colormode"}

\item{r}{shiny::reactiveValues object}

\item{verbose}{logical, diagnostic message, default FALSE}
}
\value{
used for its side effects (in particular r)
}
\description{
The server follows the "petite r" approach. It expects a reactiveValues parameter r. It modifies the "fonts" entries.
}
\details{
To save preferences across sessions, the lstore module is used if it is available.

The module follows a singleton design pattern, hence the id is preset to "fonts". It is strongly recommended to keep that id.

Run shiny::runApp(system.file("examples/03_fonts", package="shiny.gems")) to see the module in action.
}
