% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstore_srv.R
\name{lstore_srv}
\alias{lstore_srv}
\title{lstore_ui/srv is a shiny module saving/restoring data in the browser's local store}
\usage{
lstore_srv(id = "lstore", r, verbose = TRUE)
}
\arguments{
\item{id}{character, shiny id. Default "lstore"}

\item{r}{shiny::reactiveValues object}

\item{verbose}{logical, diagnostic message, default FALSE}
}
\value{
used for its side effects
}
\description{
The server follows the "petite r" approach. It expects a reactiveValues parameter r. It modifies "lstore" entries
}
\details{
The module follows a singleton design pattern, hence the id is preset to "lstore". It is strongly recommended to keep that id.

See 
shiny::runApp(system.file("examples/02_lstore", package="shiny.gems"))
to see the module in action, see colormode_srv for implementation details.
}
