% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstore_ui.R
\name{lstore_ui}
\alias{lstore_ui}
\title{lstore_ui/srv is a shiny module saving/restoring data in the browser's local store}
\usage{
lstore_ui(id)
}
\arguments{
\item{id}{character, shiny id. Default "lstore"}
}
\value{
hidden div with initialization script only
}
\description{
The UI is hidden and contains an initializiation javascript function.
}
\details{
See 
shiny::runApp(system.file("examples/02_lstore", package="shiny.gems"))
to see the module in action, see colormode_srv for implementation details.
}
