% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormode_srv.R
\name{colormode_srv}
\alias{colormode_srv}
\title{colormode_ui/srv is a shiny module for managing colors; in particular enabling dark mode.}
\usage{
colormode_srv(id = "colormode", r)
}
\arguments{
\item{id}{character, shiny id. Default "colormode"}

\item{r}{shiny::reactiveValues object}
}
\value{
the output of shiny::radioButtons()
}
\description{
The server follows the "petite r" approach. It expects a reactiveValues parameter r. It modifies entries of the "colormode
}
\details{
Currently, it is not possible to save the setting across session. This would require a user management.

The module follows a singleton design pattern, hence the id is preset to "colormode". It is strongly recommended to keep that id.

See colormode_demo to see the module in action, see colormode_srv for implementation details.
}
